Fix inconsistent toplist caching; refactor toplist and watchlist to separate helper.
